Add backend for listing update

+++ Controllers/ListingController.cs
@@ -26,11 +26,12 @@ namespace Karma.Controllers
 
         [HttpPost]
         [Authorize]
-        public ActionResult Post(Listing listing) {
+        public ActionResult Post(Listing listing)
+        {
             var random = new Random();
             listing.Id = random.Next(9999).ToString(); // temp fix for id generation, later this should be assigned in DB.
             listing.DatePublished = DateTime.UtcNow; //temp fix for curr date with form submit
-            
+
             string userId = User.FindFirst(ClaimTypes.Name)?.Value;
             listing.OwnerId = userId;
             _listingRepository.AddListing(listing);
@@ -44,15 +45,37 @@ namespace Karma.Controllers
         }
 
         [HttpGet("{id}")]
-        public Listing GetListingById(string id) 
+        public Listing GetListingById(string id)
         {
             return _listingRepository.GetListingById(id);
         }
 
         [HttpDelete("{id}")]
         [Authorize]
-        public void DeleteListing(string id) {
+        public IActionResult DeleteListing(string id)
+        {
+            string userId = User.FindFirst(ClaimTypes.Name)?.Value;
+            var listing = _listingRepository.GetListingById(id);
+            if (listing.OwnerId != userId)
+                return Unauthorized();
             _listingRepository.DeleteListingById(id);
+            return Ok();
+        }
+
+        [HttpPut("{id}")]
+        [Authorize]
+        public IActionResult UpdateListing(string id, [FromBody] Listing listing)
+        {
+
+            var old = _listingRepository.GetListingById(listing.Id);
+            if (old == null) return NotFound();
+
+            string userId = User.FindFirst(ClaimTypes.Name)?.Value;
+            if (old.OwnerId != userId) return Unauthorized();
+
+
+            _listingRepository.UpdateListing(listing);
+            return Ok();
         }
     }
 }
